requirements.txt
requests
pytest
pydantic
allure-pytest
Jinja2
xmlschema